Update for Elixir 1.2.1
Prior to this commit, LoggerFileBackend accepted any Elixir version between 1.0.0 and 1.2.0; however, 1.2.1 was just released and fails this check, as would any further future release. This commit changes the acceptable version to `~> 1.0`, which will expands to `>= 1.0 and < 2.0`. Since Elixir follows SemVer, we should should work with any future release up to a potentially backwards incompatible 2.0.