fix: warnings about the use of deprecated :warn
Convert `warn` to `warning` when a system is running Elixir >= 1.11 to
be backwards compatible wile converting the backwards compatible
value.

In Elixir 1.11 `Logger.warn` was deprecated in favor of
`Logger.warning` and logs a warning when `warn` is used. To stay
backward compatible Logger itself reports level `warning` as `warn` to
backends[^1].

When a backend passed `warn` to `Logger`, essentially feeding
the module it' own food, Logger will issue a warning about the use of
`warn`[^2].

[^1]: https://hexdocs.pm/logger/1.11.3/Logger.html#warning/2
[^2]: https://github.com/elixir-lang/elixir/blob/ad4c31eef1af836c3c9972f67b5deb6b0a541c53/lib/logger/lib/logger.ex#L1171